+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Rails 6.0.0.rc1
 * Add `EveOnline::ESI::CharacterCalendarEvent`
 * Add `EveOnline::ESI::Models::EventDetails`
+* Add `EveOnline::ESI::Models::LoyaltyStoreOffer`
 
 ## v0.26.0